Rename target function transformer
